+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release date `UNRELEASED`
 - [ test ] - Add ring response tests.
 - [ session ] - Overwrite ring session data accordingly to spec.
 - [ test ] - Add ring session tests.
+- [ session ] - Handle :recreate meta, unbind session on delete.
 
 ## `1.0.65-beta10`